requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.12.0
asgiref==3.11.0
asttokens==3.0.1
backoff==2.2.1
billiard==4.2.4
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
decorator==5.2.1
distro==1.9.0
dj-database-url==3.0.1
Django==4.2.10
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-extensions==3.2.3
django-redis==5.4.0
django-timezone-field==7.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
executing==2.2.1
filelock==3.20.1
fsspec==2025.12.0
gunicorn==21.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.27.0
huggingface_hub==1.2.3
idna==3.11
ipython==8.20.0
jedi==0.19.2
jiter==0.12.0
joblib==1.5.3
kombu==5.6.1
matplotlib-inline==0.2.1
monotonic==1.6
numpy==2.4.0
openai==1.54.0
packaging==25.0
parso==0.8.5
posthog==3.1.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
psycopg2-binary==2.9.11
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.3
redis==5.0.1
requests==2.31.0
scikit-learn==1.8.0
scipy==1.16.3
sentry-sdk==1.40.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
stack-data==0.6.3
threadpoolctl==3.6.0
tokenizers==0.22.1
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.6.0
